Extract ThemeChooser component
